# lib 2/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Personal Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  completed: 'completed',
  missed: 'missed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  dueTime?: string | null;
  /** @nullable */
  tagId?: number | null;
  /** @nullable */
  tagName?: string | null;
  /** @nullable */
  tagColor?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  fromPdf?: boolean;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority?: TaskInputPriority;
  dueDate?: string;
  dueTime?: string;
  tagId?: number;
  fromPdf?: boolean;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
  missed: 'missed',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: TaskUpdatePriority;
  dueDate?: string;
  dueTime?: string;
  tagId?: number;
  status?: TaskUpdateStatus;
}

export interface RescheduleInput {
  dueDate: string;
  dueTime?: string;
  note?: string;
}

export interface TaskSummary {
  total: number;
  completed: number;
  pending: number;
  missed: number;
  completionRate?: number;
}

export interface Habit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  color?: string | null;
  streak: number;
  currentStreak: number;
  longestStreak?: number;
  level: number;
  completedToday?: boolean;
  /** @nullable */
  lastChecked?: string | null;
  /** @nullable */
  tagId?: number | null;
  createdAt: string;
}

export interface HabitInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  icon?: string;
  color?: string;
  tagId?: number;
}

export interface HabitUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  icon?: string;
  color?: string;
  tagId?: number;
}

export type FinanceEntryType = typeof FinanceEntryType[keyof typeof FinanceEntryType];


export const FinanceEntryType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceEntry {
  id: number;
  type: FinanceEntryType;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  date: string;
  /** @nullable */
  dueDate?: string | null;
  isPaid?: boolean;
  /** @nullable */
  tagId?: number | null;
  createdAt: string;
}

export type FinanceInputType = typeof FinanceInputType[keyof typeof FinanceInputType];


export const FinanceInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceInput {
  type: FinanceInputType;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  description?: string;
  date: string;
  dueDate?: string;
  isPaid?: boolean;
  tagId?: number;
}

export type FinanceUpdateType = typeof FinanceUpdateType[keyof typeof FinanceUpdateType];


export const FinanceUpdateType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceUpdate {
  type?: FinanceUpdateType;
  /** @minimum 0.01 */
  amount?: number;
  category?: string;
  description?: string;
  date?: string;
  dueDate?: string;
  isPaid?: boolean;
  tagId?: number;
}

export interface CategoryBreakdown {
  category: string;
  total: number;
  count: number;
  percentage?: number;
}

export interface MonthlySummary {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netBalance: number;
  expensesByCategory: CategoryBreakdown[];
  incomeByCategory?: CategoryBreakdown[];
}

export interface Tag {
  id: number;
  name: string;
  color: string;
  /** @nullable */
  icon?: string | null;
  isCustom: boolean;
  createdAt: string;
}

export interface TagInput {
  /** @minLength 1 */
  name: string;
  color: string;
  icon?: string;
}

export interface UserStats {
  dailyStreak: number;
  level: number;
  xp: number;
  xpToNextLevel: number;
  totalTasksCompleted: number;
  totalHabitsChecked: number;
  badges: number;
  weeklyCompletionRate?: number;
  streakBroken?: boolean;
  allTasksCompletedToday?: boolean;
}

export type BadgeRarity = typeof BadgeRarity[keyof typeof BadgeRarity];


export const BadgeRarity = {
  common: 'common',
  rare: 'rare',
  epic: 'epic',
  legendary: 'legendary',
} as const;

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  earnedAt: string;
  rarity?: BadgeRarity;
}

export interface DailyQuote {
  quote: string;
  author: string;
  date?: string;
}

export interface PdfTextInput {
  /** @minLength 1 */
  text: string;
}

export type PdfTaskSuggestionPriority = typeof PdfTaskSuggestionPriority[keyof typeof PdfTaskSuggestionPriority];


export const PdfTaskSuggestionPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface PdfTaskSuggestion {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: PdfTaskSuggestionPriority;
  confidence: number;
}

export interface PdfTaskConfirmInput {
  tasks: TaskInput[];
}

export type ListTasksParams = {
status?: ListTasksStatus;
/**
 * Filter by due date (YYYY-MM-DD)
 */
date?: string;
/**
 * @nullable
 */
tagId?: number | null;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  pending: 'pending',
  completed: 'completed',
  missed: 'missed',
  all: 'all',
} as const;

export type ListFinancesParams = {
type?: ListFinancesType;
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
category?: string;
};

export type ListFinancesType = typeof ListFinancesType[keyof typeof ListFinancesType];


export const ListFinancesType = {
  income: 'income',
  expense: 'expense',
  all: 'all',
} as const;

export type GetMonthlySummaryParams = {
/**
 * Month in YYYY-MM format
 */
month: string;
};