# lib 2/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Productivity Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finances:
    get:
      operationId: listFinances
      tags: [finances]
      summary: List financial entries
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, all]
        - name: month
          in: query
          schema:
            type: string
            description: Filter by month (YYYY-MM)
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of financial entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceEntry"
    post:
      operationId: createFinanceEntry
      tags: [finances]
      summary: Create a financial entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceEntry"
  /finances/{id}:
    patch:
      operationId: updateFinanceEntry
      tags: [finances]
      summary: Update a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceEntry"
    delete:
      operationId: deleteFinanceEntry
      tags: [finances]
      summary: Delete a financial entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /finances/monthly-summary:
    get:
      operationId: getMonthlySummary
      tags: [finances]
      summary: Get monthly spending breakdown by category
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            description: Month in YYYY-MM format
      responses:
        "200":
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySummary"

  # ── TAGS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FinanceEntry:
      type: object
      required: [id, type, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        dueDate:
          type: ["string", "null"]
        isPaid:
          type: boolean
        tagId:
          type: ["integer", "null"]
        createdAt:
          type: string
    FinanceInput:
      type: object
      required: [type, amount, category, date]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
        dueDate:
          type: string
        isPaid:
          type: boolean
        tagId:
          type: integer
    FinanceUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
        description:
          type: string
        date:
          type: string
        dueDate:
          type: string
        isPaid:
          type: boolean
        tagId:
          type: integer
    CategoryBreakdown:
      type: object
      required: [category, total, count]
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
    MonthlySummary:
      type: object
      required: [month, totalIncome, totalExpenses, netBalance, expensesByCategory]
      properties:
        month:
          type: string
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netBalance:
          type: number
        expensesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        incomeByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"

    # ── Tag schemas ───────────────────────────────────────────────────────────
...